ExpressionParser: keep track of parsing state
To support unary operators, it is crucial to know exactly where we are
within the infix expression. In particular, there are two common states:

A) immediately following an infix binary operator; and
B) not immediately following such an operator.

After we see an infix binary operator, we enter state A; in other cases,
we enter state B. Keeping track of this state maintains the context
needed to decide whether each sort of operator is appropriate for that
specific circumstance. For example, it is illegal to have two infix
binary operators in a row (e.g., "a++b").

In order to make the code (hopefully) easier to understand, we actually
maintain _three_ state flags: PREFIX_OK, POSTFIX_OK and INFIX_OK, which
describe whether it would be legal to encounter a following prefix unary
operator, postfix unary operator or infix binary operator, respectively.
However, as described above, there are really only two states:

A) PREFIX_OK; and
B) POSTFIX_OK | INFIX_OK.

But we'll leave the distinction in the codebase in case we ever add
support for a hypothetical more esoteric case or operator.