ExpressionParser: be strict about accepting tokens
There are certain kinds of tokens that are not acceptable in certain
contexts. In particular, if we are expecting an infix (or postfix)
operator, there are many kinds of tokens we should _not_ accept,
including literals, variables and functions.

We should also only accept commas when we would accept an infix
operator. After all, the comma _is_ a sort of infix operator for
separating function arguments. If we were really clever, we would
always treat it that way, with the comma operator concatenating
list items, and all functions being unary and operating on lists.
But I'm not sure I want to be that clever at this moment in time.