ExpressionParserTest: tweak unary tests
The tests were intended to verify unary vs. binary minus in particular.
But we are about to add '--' to the list of standard operators, which
will hose up these tests, since the '--' operators have higher
precedence than both unary and binary minus.

So let's change the tests to alternate between minuses and pluses. This
should be OK since '-+' and '+-' are not planned as standard operators.