StandardStackEvaluator: use consistent ordering
Most operations declare their helper methods in the order int < long <
float < double < BigInteger < BigDecimal, but there were a couple of
outliers. This updates things to be consistent in that regard.

The reason for that order is that it represents Java's widening
assignment order (with partial extension to BigInteger and BigDecimal).