Split out some of DefaultEvaluator's logic
The standard operator methods belong in an interface: StandardEvaluator.

The variable-centric -- but type-agnostic -- logic of preInc, preDec,
postInc, postDec, assign, and assign* belong in an abstract base class,
AbstractStandardStackEvaluator, along with the StackEvaluator#execute
implementation which delegates to each of the standard operator methods.

That leaves DefaultEvaluator with its sole responsibility being
implementation of the value-centric (i.e., not assignment-centric)
standard operators for the common basic types.