Add support for groups, beyond only parentheses
A group is an N-ary operator noted with a left-hand initiator symbol and
a right-hand terminator symbol. Parentheses are one example of a group,
but so are square brackets, as well as curly braces. This change adds
support for all three such groups.

This change reworks the meaning of Function; it is now always a binary
operator which indicates an implicit operator between a noun (typically,
but not always, a variable) and a group -- e.g.: f(1, 2, 3).

As such, the Verb base class is now no longer necessary, since all verbs
are now operators: we have the vanilla Operator class, and its two
subclasses Function and Group. Therefore, this commit squashes the
general Verb functionality into the Operator base class directly.

This is a backwards incompatible change, as it alters the contents of
parsed postfix queues and syntax trees. So we bump to 3.0.0-SNAPSHOT.