Provide a hook for operators to clone themselves
When adding an operator to the stackd during parsing,
we might want to add a copy of that operator... or we might not.
The instance() method provides a seam for Operator subclasses to
decide that for themselves. This will come in handy shortly,
when we add more flexible support for groups.