DefaultEvaluator: support 1-D array/list access
This is a basic implementation which (for now) only supports Lists,
and does not care which grouping operator was used to create them.

Ultimately, the different grouping operators could produce different
flavors of collection data structures, and the function implementation
could respond differently to them. But this is good enough for now.