Evaluator: provide an accessor for the parser used
Previously, there was no way of directly accessing the ExpressionParser
being used internally by the Evaluator.

While I can imagine an Evaluator implementation that is not backed by
an ExpressionParser, it would be really odd in the context of SJEP,
so let's not worry too much about that theoretical case. (And in that
scenario, getParser() could just return null anyway.)