EvaluatorConsole: improve output of postfix queues
The evaluator was not able to print the result of the postfix built-in
function successfully when variables were involved, because it tried to
query variable values in strict mode. The avoid this issue, the
evaluator now uses non-strict mode when printing results.

We now also print the type of each element of the queue, for clarity.