Unify logic for standard operators on common types
It was cut-and-pasted from DefaultStackEvaluator to
DefaultTreeEvaluator. But it can go into an abstract base class.

I thought about making a StandardCommonTypesEvaluator interface,
and putting this logic in there, but it would have made public many
of the private helper methods currently in use. So for now, we keep
the logic as an abstract base class.